refactor: product api 정리(LUP-119)
