--- frontend/src/apis/getCartList.ts
@@ -1,12 +1,0 @@
-import axios from 'axios';
-import { CartType } from '@/assets/types/CartType';
-import client from './client';
-
-interface cartRequest {
-  carts : CartType[]
-}
-
-export const getCartList = async (): Promise<CartType[]> => {
-  const { data } = await client.get<cartRequest>('/cart');
-  return data.carts;
-};
